+++ contracts/src/verifiers/WithdrawLocalGroth16Verifier.sol
@@ -25,17 +25,17 @@ pragma solidity >=0.7.0 <0.9.0;
 
 contract WithdrawLocalGroth16Verifier {
     // Scalar field size
-    uint256 constant r    = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
+    uint256 constant r = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
     // Base field size
-    uint256 constant q   = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
+    uint256 constant q = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
 
     // Verification Key data
-    uint256 constant alphax  = 10314683402145919335415264089338013869151872735661243528435829273762895412475;
-    uint256 constant alphay  = 15311410802386913807174485311770598542990692773058369166097347676916826427424;
-    uint256 constant betax1  = 8742476040979126669529512667270167370972734897784767815803986702043271844587;
-    uint256 constant betax2  = 21546977338313367449764778081974431327514198880463503495700342085596869133184;
-    uint256 constant betay1  = 9289809140465907199790286310535739942219483053026485452107542051245760617369;
-    uint256 constant betay2  = 10645240371221413259645920038475973272479052146228783657329287031421083633145;
+    uint256 constant alphax = 10314683402145919335415264089338013869151872735661243528435829273762895412475;
+    uint256 constant alphay = 15311410802386913807174485311770598542990692773058369166097347676916826427424;
+    uint256 constant betax1 = 8742476040979126669529512667270167370972734897784767815803986702043271844587;
+    uint256 constant betax2 = 21546977338313367449764778081974431327514198880463503495700342085596869133184;
+    uint256 constant betay1 = 9289809140465907199790286310535739942219483053026485452107542051245760617369;
+    uint256 constant betay2 = 10645240371221413259645920038475973272479052146228783657329287031421083633145;
     uint256 constant gammax1 = 19786488175694835941529082486176010093671788452701050737945948286615958246569;
     uint256 constant gammax2 = 20330634461338860209244322586166193708999379153762374339780730602203574324967;
     uint256 constant gammay1 = 5928072313096986966179943778308363878908435360419431578510684278352903202909;
@@ -45,27 +45,30 @@ contract WithdrawLocalGroth16Verifier {
     uint256 constant deltay1 = 8810824702403597757041960647441338188287452861929357829520560151016180636116;
     uint256 constant deltay2 = 20423055301853288990931997049485455849638502515255059322630657921443113246087;
 
-    
     uint256 constant IC0x = 5981170191792726018550786071194620221508954811380977324254400290854175013253;
     uint256 constant IC0y = 7963352545678533703769588588119136330774730818542667732095523856920941509207;
-    
+
     uint256 constant IC1x = 20821657135676432889142208462570017708140579915257361314167309290885899903208;
     uint256 constant IC1y = 11644784712562095374763108262330308260517779877916458745595529801600526776721;
-    
+
     uint256 constant IC2x = 1851457957886242515550697099551094984872412297309215389748394026781243764292;
     uint256 constant IC2y = 18466285845908517752405622355001390565150981176040526649859134446727362915453;
-    
+
     uint256 constant IC3x = 3452574657221297304286562511070594155387565234735763151066981242551916564841;
     uint256 constant IC3y = 17620257701175947555528870367770816985691621663381315377899926759899247136105;
-    
-    
+
     // Memory data
     uint16 constant pVk = 0;
     uint16 constant pPairing = 128;
 
     uint16 constant pLastMem = 896;
 
-    function verifyProof(uint[2] calldata _pA, uint[2][2] calldata _pB, uint[2] calldata _pC, uint[3] calldata _pubSignals) public view returns (bool) {
+    function verifyProof(
+        uint256[2] calldata _pA,
+        uint256[2][2] calldata _pB,
+        uint256[2] calldata _pC,
+        uint256[3] calldata _pubSignals
+    ) public view returns (bool) {
         assembly {
             function checkField(v) {
                 if iszero(lt(v, r)) {
@@ -73,7 +76,7 @@ contract WithdrawLocalGroth16Verifier {
                     return(0, 0x20)
                 }
             }
-            
+
             // G1 function to multiply a G1 value(x,y) to value in an address
             function g1_mulAccC(pR, x, y, s) {
                 let success
@@ -108,8 +111,7 @@ contract WithdrawLocalGroth16Verifier {
                 mstore(add(_pVk, 32), IC0y)
 
                 // Compute the linear combination vk_x
-                
-                
+
                 g1_mulAccC(_pVk, IC1x, IC1y, calldataload(add(pubSignals, 0)))
                 g1_mulAccC(_pVk, IC2x, IC2y, calldataload(add(pubSignals, 32)))
                 g1_mulAccC(_pVk, IC3x, IC3y, calldataload(add(pubSignals, 64)))
@@ -138,7 +140,6 @@ contract WithdrawLocalGroth16Verifier {
                 mstore(add(_pPairing, 384), mload(add(pMem, pVk)))
                 mstore(add(_pPairing, 416), mload(add(pMem, add(pVk, 32))))
 
-
                 // gamma2
                 mstore(add(_pPairing, 448), gammax1)
                 mstore(add(_pPairing, 480), gammax2)
@@ -155,7 +156,6 @@ contract WithdrawLocalGroth16Verifier {
                 mstore(add(_pPairing, 704), deltay1)
                 mstore(add(_pPairing, 736), deltay2)
 
-
                 let success := staticcall(sub(gas(), 2000), 8, _pPairing, 768, _pPairing, 0x20)
 
                 isOk := and(success, mload(_pPairing))
@@ -165,21 +165,20 @@ contract WithdrawLocalGroth16Verifier {
             mstore(0x40, add(pMem, pLastMem))
 
             // Validate that all evaluations ∈ F
-            
+
             checkField(calldataload(add(_pubSignals, 0)))
-            
+
             checkField(calldataload(add(_pubSignals, 32)))
-            
+
             checkField(calldataload(add(_pubSignals, 64)))
-            
+
             checkField(calldataload(add(_pubSignals, 96)))
-            
 
             // Validate all evaluations
             let isValid := checkPairing(_pA, _pB, _pC, _pubSignals, pMem)
 
             mstore(0, isValid)
-            
+
             return(0, 0x20)
         }
     }
